requirements.txt
litellm
fastapi
uvicorn[standard]

# observability/tracing dependencies
arize-phoenix
openinference-instrumentation-openai
opentelemetry-exporter-otlp
